In embed.pl, refactor readvars() to return a sorted list.
Previously it would assign to the hash reference passed to it. However, the
rest of the code only ever used the keys of that hash, in sorted order, so it
is simpler to change readvars() to return that list directly.

Also move the declaration of the lexical $sym down to its first point of use.