podcheck.t: Change verbatim line length max to 79
Some pagers reserve a column for their own use.  This patch has
podcheck.t warn if a verbatim line will wrap when such a pager is used
in an 80 column terminal window.  This effectively restores the
situation to what it was before commit
2cd46bfd6d03cd8826c5e968efdd36416191bc8f, but the message is now honest
about the situation, i.e., that 79 colummns are effectively the max
without a warning.

The point of this is to give a number to shoot for, 79, that will work
almost always (I believe though that there are pagers that reserve 2
columns in some cases), but of course the data file is continued to be
used to silence the warning if it isn't desired to get down to 79.