Change core_prototype to take a keyword num
This refactoring requires the caller to provide the keyword
number to core_prototype.  Consequently, it speeds up the code in
gv.c:gv_fetchpvn_flags by allowing it to avoid an extra call to
keyword().

This takes the place of the len parameter, which is no longer used.
It used to be used only as an argument to keyword().  Since the code
that uses strEQ is only reached if the keyword has already been veri-
fied by keyword(), the name simply cannot have embedded nulls, so len
is not necessary.