API to explicitly select to export XSUB symbols or not
This adds a few additional macros to XSUB.h:
XS_INTERNAL and XS_EXTERNAL are versions of the XS macro
that explicitly use internal (static) linking or not.
XSPROTO_INTERNAL and XSPROTO_EXTERNAL are the obvious equivalents
for XSPROTO (which is apparently not public yet we support its
use in SWIG...).

The XS and XSPROTO macros themselves are not defined to not
use STATIC, but this may (should?) be changed in the future.