Change (l)stat’s prototype from * to ;*
Since the argument is optional, we need a semicolon.

This commit accomplishes that by setting the OA_OPTIONAL flag for the
appropriate  entries in PL_opargs.  This should not affect anything
else, as ck_ftst (the check routine for [l]stat) doesn’t even look at
PL_opargs.

It also has to tweak the prototype-generation logic slightly, because
PL_opargs also has OA_DEFGV set.

I think this is insignificant enough not to warrant a delta entry.