Remove select’s prototype
select has a prototype of (;*), which is incorrect, as it implied that
it has high precedence.  It also fails to account for the four-argu-
ment form.  While removing all incorrect prototypes is counterproduc-
tive, I think this one is wrong enough it deserves to go.  (And the
precedence problem is a good argument against it, as there is cur-
rently no other way to set precedence.)