Trim trailing dots on VMS in podcheck.t.
readdir() and thus File::Find return files having zero-length
extensions as 'foo.' which doesn't match entries like 'foo' in
the known problems database.  Trimming that trailing dot reduces
the number of failures on VMS to a handful.  They are primarily
due to:

1.)  The test attempts to exclude .PL files, but on VMS, case is
not preserved by default, so these appear as .pl and are not
excluded, confusing a number of the tests.

2.)  Utilities and scripts are built with a .com extension on VMS,
but the corresponding entries in the known problems database have
no extension, so they don't match.