Initial attempt at putting podcheck.t on a diet.
The current design of this test is to scan *every* file in a dirty
build directory to see if it contains POD, excluding only a short
list of files and file types that it knows about.  This means that
on most platforms it was scanning executable code, object code,
object libraries, dynamic libraries, and many other build products
too numerous to mention.

The present commit expands the exclusion list considerably and
makes that list slightly more cross-platform, but it's probably
still quite incomplete.  Whether that list is even worth
maintaining or the test should be based on an inclusion list
rather than an exclusion list is a question worth asking.

On my system, the present austerity measures, despite the addition
of Config, reduce memory consumption by 14% (but it's still a pig,
gulping something like 160MB).  Physical I/O is reduced by 105%
and elapsed time by 25%.