mktables: use full_name as well when converting table to binary
This fixes a bug in which a table gets converted to binary, but there is
no distinguishing between the regular and full name for it.

There is no test for this, as it only shows up if an installation
chooses to download and compile the Unicode Unihan data base, something
that isn't shipped by default.