mktables: Add defensive coding
$b should always be defined here, since earlier we tested that the
range counts were the same between $self and $other.  Yet, in a
configuration of this program not committed, the range counts were the
same, but $other->ranges generated an empty list.  I believe this to be
a core bug, but I was unable to narrow it down, and so added this
defensive coding measure.