mktables: Try harder to find a good table name
The short name of a table is used to generate the file name in which it
is to be stored.  If the short name is a null string, the file is given
the name of '_.pl'.  A more meaningful name would be preferred.  If the
table is equivalent to other tables, we can use their name instead of
its.  This commit does that.