Make unlink with implicit $_ work once more
I broke this with commit ea5703f4.  unlink is the only op that has the
OA_DEFGV flag and no scalar or file arguments.

Commit ea5703f4 changed the OA_DEFGV logic in ck_fun to generate a
new $_ op for the first optional parameter, instead of just the first
parameter, to avoid special-casing unpack.  But lists are not marked
with OA_OPTIONAL.  So this commit changes it to check for list parame-
ters as well.