Get coreamp.t passing on VMS.
The last record in a file ends with a newline willy nilly, so we
might as well write (and test for) one explicitly.

The name of a directory file ends with .DIR, so we need to allow
for that.

File::Temp panics when trying to delete the directory that is the
current working directory, so save where we were and restore to it
before clean-up.