Close some unclosed comments in vms/vms.c.
Clearly this code is not well-tested.  Especially the parts that
the compiler couldn't see because it was hidden by unclosed
comment blocks.  Ouch.

I'm not going to name names or look too deep in the blame log, but
suffice it to say that this is why patches that add hundreds of
lines of new code without tests should not be accepted.