Unicode::UCD: speed up some look ups
A previous commit created the possibility that for a data structure to
be used instead of an array of arrays.  The latter data structure is
used because most properties have long ranges in which each code point
maps to the same thing, and so don't have to have a hash entry for each
code point, can use the range as the base data structure.  However
certain properties (or property-like structures) don't have long ranges,
and hence don't need to save memory by having things work through ranges
instead of individual code points.  For these, a hash offers improved
speed, without the memory cost.

This patch converts 4 of the data structures to hashes.  Future
extensions to this module will also take advantage of some of these
being in hashes